initial sync refactoring
